Very mildly reorganized the text/example stuff
