+++ src/test/kotlin/dtx/example/rs_tables/PlayerExtensions.kt
@@ -1,0 +1,7 @@
+package dtx.example.rs_tables
+
+import dtx.example.Player
+
+fun Player.sendMessage(message: String) {
+    println("[to: $username]: $message")
+}
